requirements.txt
asgiref==3.7.2
boto3==1.28.4
botocore==1.31.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coverage==7.2.7
cryptography==41.0.4
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.3
django-cors-headers==4.2.0
django-countries==7.5.1
django-filter==23.2
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.1.0
django-rest-passwordreset==1.3.0
django-ses==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.6
ecdsa==0.18.0
exceptiongroup==1.1.2
Faker==19.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.17
Pillow==10.0.0
pluggy==1.2.0
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
pytest==7.4.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
